pyproject.toml
[project]
name = "fastapi-authorization-gateway"
description = "A route-based authorization framework for FastAPI"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "Edward Keeble", email = "ed@developmentseed.org"},
    {name = "Anthony Lukach", email = "anthony@developmentseed.org"},
    {name = "Vincent Sarago", email = "vincent@developmentseed.org"}
]
keywords = ["FastAPI", "Authorization"]
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = ["version"]
dependencies = [
    "fastapi-slim>=0.111.0",
    "typing_extensions; python_version < '3.9'",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "httpx",
]
dev = [
    "pre-commit",
]

[project.urls]
Homepage = "https://github.com/developmentseed/fastapi-authorization-gateway"
Source = "https://github.com/developmentseed/fastapi-authorization-gateway"

[tool.hatch.version]
path = "fastapi_authorization_gateway/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = [
  "/tests",
  "/docs",
  ".pytest_cache",
  ".github",
  ".gitignore",
  ".pre-commit-config.yaml",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.run]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.isort]
profile = "black"
known_first_party = ["tipg"]
known_third_party = ["geojson_pydantic", "buildpg", "pydantic"]
forced_separate = [
    "fastapi",
    "starlette",
]
default_section = "THIRDPARTY"

[tool.mypy]
no_strict_optional = "True"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # flake8
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B905",  # ignore zip() without an explicit strict= parameter, only support with python >3.10
]
